Hide typesafe wrapper class inside source generator internal namespace
