--- Packages/com.unity.inputsystem/InputSystem/TypeSafeAPI.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d6bf60707bafde34caccb746fde88430
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
