--- Packages/com.unity.inputsystem/InputSystem/TypeSafeAPI/Input.actions.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ecaba2a1b6fed3a499f1da89e597b446
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
